requirements.txt
httpx>=0.24.1
readability-lxml>=0.8.1
beautifulsoup4>=4.12.2
openpyxl>=3.1.2
pyinstaller>=6.0.0 